requirements.txt
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
django-ckeditor-5==0.2.18 ; python_version >= "3.11" and python_version < "4.0"
django-js-asset==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
django-mptt-admin==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
django-mptt==0.17.0 ; python_version >= "3.11" and python_version < "4.0"
django-parler-rest==2.2 ; python_version >= "3.11" and python_version < "4.0"
django-parler==2.3 ; python_version >= "3.11" and python_version < "4.0"
django-reversion==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
django-taggit-autosuggest==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
django-taggit==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
django==5.2.3 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.16.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.2.1 ; python_version >= "3.11" and python_version < "4.0"
psycopg-binary==3.2.9 ; implementation_name != "pypy" and python_version >= "3.11" and python_version < "4.0"
psycopg[binary]==3.2.9 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.11" and python_version < "3.13"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
uvicorn==0.34.3 ; python_version >= "3.11" and python_version < "4.0"
